gpiote: protect against sense listener chain bugs
Correct sense_listener constructor to retain unlinked status of newly
created instance.  Prevent corrupting the chain by enable or disable
operations on an invalid state.

+++ src/gpiote.cc
@@ -216,14 +216,18 @@ void
 GPIOTE::sense_listener::enable ()
 {
   mutex_type mutex;
-  sense_chain_.link_back(*this);
+  if (sense_chain_.is_unlinked(*this)) {
+    sense_chain_.link_back(*this);
+  }
 }
 
 void
 GPIOTE::sense_listener::disable ()
 {
   mutex_type mutex;
-  sense_chain_.unlink(*this);
+  if (!sense_chain_.is_unlinked(*this)) {
+    sense_chain_.unlink(*this);
+  }
 }
 
 void
